Setting up gulp tasks and remove unnecessary scripts.
We needed to rearrange the directory structure to get gulp to build the site in an acceptable way. At this point gulp is handling a few things for us: compiling our sass, concatenating and minifying our javascript, and building the jekyll site. We deleted more unneeded scripts that were causing problems.