Replacing categories with list.js search.
We used to group recipes by arbitrarily set categories. Not only was this brittle, it wasn't exactly the most helpful. Now we have a simple search bar that finds recipes that include the searched text. This is far better since it allows users to  more reliably find what they are looking for.